requirements.txt
absl-py==2.1.0
accelerate==1.5.2
acres==0.3.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bioc==2.1
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
boto3==1.37.13
botocore==1.37.13
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
cloudpathlib==0.20.0
cloudpickle==3.1.1
colorama==0.4.6
confection==0.1.5
configobj==5.0.9
configparser==7.2.0
conllu==4.5.3
contourpy==1.3.1
cryptography==44.0.2
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
dask==2025.2.0
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
diff-match-patch==20241021
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
docopt==0.6.2
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
en_core_web_trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.8.0/en_core_web_trf-3.8.0-py3-none-any.whl#sha256=272a31e9d8530d1e075351d30a462d7e80e31da23574f1b274e200f3fff35bf5
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.10
fastjsonschema==2.21.1
filelock==3.17.0
fitz==0.0.1.dev2
flair==0.15.1
Flask==3.1.0
flask-cors==5.0.1
Flask-WTF==1.2.2
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.2.0
ftfy==6.3.1
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20.3
greenlet==3.1.1
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
intervaltree==3.1.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyterlab_pygments==0.3.0
keras==3.9.0
Keras-Preprocessing==1.1.2
keybert==0.9.0
kiwisolver==1.4.8
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libclang==18.1.1
locket==1.0.0
looseversion==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.4.0
more-itertools==10.6.0
mpld3==0.5.10
mpmath==1.3.0
murmurhash==1.0.12
mysqlclient==2.2.7
namex==0.0.8
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.66.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pathlib==1.0.1
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.5
pickle-mixin==1.0.2
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pptree==3.1
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.3
prov==2.0.1
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
puremagic==1.28
py4j==0.10.9.7
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydot==3.0.4
Pygments==2.19.1
PyMuPDF==1.25.3
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
PySocks==1.7.1
pyspark==3.5.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-json-logger==3.2.1
pytorch_revgrad==0.2.0
pytz==2025.1
pywin32==308
pywinpty==2.0.15
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==6.3.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
segtok==1.5.11
Send2Trash==1.8.3
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
sqlitedict==2.1.0
sqlparse==0.5.3
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.0
streamlit==1.43.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_cpu==2.19.0
tensorflow_intel==2.18.0
termcolor==2.5.0
terminado==0.18.1
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
transformer-smaller-training-vocab==0.4.0
transformers==4.49.0
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2025.1
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
watchdog==6.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
Wikipedia-API==0.8.1
wrapt==1.17.2
WTForms==3.2.1
zipp==3.21.0
